+++ changelog.md
@@ -10,6 +10,8 @@
   ([#108](https://github.com/feltcoop/felt/pull/108))
 - publish `ui/Message.svelte`
   ([#108](https://github.com/feltcoop/felt/pull/108))
+- add vendored `util/typebox.ts`
+  ([#113](https://github.com/feltcoop/felt/pull/113))
 
 ## 0.5.5