+++ changelog.md
@@ -4,6 +4,8 @@
 
 - **break**: rename and add exports in `ui/icons.ts`
   ([#108](https://github.com/feltcoop/felt/pull/108))
+- **break**: add peer deps `svelte` and `@sveltejs/kit`
+  ([#112](https://github.com/feltcoop/felt/pull/112))
 - update `ui/Pending_Animation.svelte` with slot content and `paused` prop
   ([#108](https://github.com/feltcoop/felt/pull/108))
 - replace `ui/Message.svelte` prop `text` with a slot